backend/streamlit_backend/requirements_streamlit.txt
# Streamlit Cloud Backend Requirements
# Smart Energy Consumption Backend API

# Core Streamlit
streamlit>=1.28.0

# Django & REST Framework
Django>=5.2.4
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
djangorestframework_simplejwt>=5.5.1

# Database & ORM
sqlparse>=0.5.3
pytz>=2025.2
tzdata>=2025.2

# Authentication & Security
PyJWT>=2.10.1
python-decouple>=3.8

# Data Processing
numpy>=2.2.6
pandas>=2.3.1
scikit-learn>=1.7.1
scipy>=1.16.1
joblib>=1.5.1

# Image Processing
opencv-python>=4.12.0.88
pillow>=11.3.0
pytesseract>=0.3.13

# HTTP & API
requests>=2.32.4
urllib3>=2.5.0
certifi>=2025.7.14
charset-normalizer>=3.4.2
idna>=3.10

# Utilities
python-dateutil>=2.9.0.post0
six>=1.17.0
threadpoolctl>=3.6.0
wheel>=0.45.1
packaging>=25.0

# Environment & Configuration
python-dotenv>=1.0.0 